+++ Ryujinx.HLE/HOS/Services/Arp/ApplicationLaunchProperty.cs
@@ -1,4 +1,6 @@
-﻿namespace Ryujinx.HLE.HOS.Services.Arp
+﻿using Ryujinx.HLE.FileSystem;
+
+namespace Ryujinx.HLE.HOS.Services.Arp
 {
     class ApplicationLaunchProperty
     {
@@ -7,5 +9,16 @@
         public byte  BaseGameStorageId;
         public byte  UpdateGameStorageId;
         public short Padding;
+
+        public static ApplicationLaunchProperty GetDefault()
+        {
+            return new ApplicationLaunchProperty
+            {
+                TitleId             = 0x00,
+                Version             = 0x00,
+                BaseGameStorageId   = (byte)StorageId.NandSystem,
+                UpdateGameStorageId = (byte)StorageId.None
+            };
+        }
     }
 }
